+++ src/widgets/TbDataColumn.php
@@ -36,9 +36,15 @@ class TbDataColumn extends CDataColumn
 			$sort = $this->grid->dataProvider->getSort();
 			$label = isset($this->header) ? $this->header : $sort->resolveLabel($this->name);
 
-			if ($sort->resolveAttribute($this->name) !== false) {
-				$label .= '<span class="caret"></span>';
-			}
+			if ($sort->resolveAttribute($this->name) !== false){
+                            if($sort->getDirection($this->name) === CSort::SORT_ASC){
+                                $label .= ' <span class="icon-sort-down"></span>';
+                            } elseif($sort->getDirection($this->name) === CSort::SORT_DESC){
+                                $label .= ' <span class="icon-sort-up"></span>';
+                            } else {
+                                $label .= ' ';
+                            }
+                        }
 
 			echo $sort->link($this->name, $label, array('class' => 'sort-link'));
 		} else {