+++ src/widgets/input/TbInputVertical.php
@@ -140,6 +140,39 @@ class TbInputVertical extends TbInput
 	}
 
 	/**
+	 * Renders a Pass*Field field.
+	 * @return string the rendered content
+	 * @author Hrumpa
+	 */
+	protected function passfieldField()
+	{
+		if (isset($this->htmlOptions['options'])) {
+			$options = $this->htmlOptions['options'];
+			unset($this->htmlOptions['options']);
+		}
+
+		if (isset($this->htmlOptions['events'])) {
+			$events = $this->htmlOptions['events'];
+			unset($this->htmlOptions['events']);
+		}
+
+		echo $this->getLabel();
+		echo $this->getPrepend();
+		$this->widget(
+			'bootstrap.widgets.TbPassfield',
+			array(
+				'model' => $this->model,
+				'attribute' => $this->attribute,
+				'options' => isset($options) ? $options : array(),
+				'events' => isset($events) ? $events : array(),
+				'htmlOptions' => $this->htmlOptions,
+			)
+		);
+		echo $this->getAppend();
+		echo $this->getError() . $this->getHint();
+	}
+
+	/**
 	 * Renders a radio button.
 	 * @return string the rendered content
 	 */
@@ -297,7 +330,7 @@ class TbInputVertical extends TbInput
 		echo $this->getAppend();
 		echo $this->getError() . $this->getHint();
 	}
-	
+
 	/**
 	 * Renders a datetimepicker field.
 	 * @return string the rendered content