+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Antonio Ramirez.
 
 
 ## YiiBooster version 2.0.0
+- **(enh)** Pass*Field by Antelle (v 1.1.7) widget added #667 (Hrumpa)
 - **(cln)** Removed long-broken `TbExtendedTooltip` from codebase #511 (hijarian)
 - **(fix)** Checkboxes, selects and radiobuttons in form rows do not generate invalid HTML markup anymore #664 #626 (ZhandosKz)
 - **(enh)** Now you can properly extend/modify the rendering of individual alerts in `TbAlert` #619 (hijarian)